+++ crates/devela/macros/src/copied/doclink.rs
@@ -16,7 +16,6 @@ pub(crate) use _DOCLINK_CUSTOM_DOMAIN;
 /// Helps doc-linking items in downstream crates.
 /// <!-- (*internet* version) -->
 #[cfg(feature = "__publish")]
-#[allow(clippy::crate_in_macro_def, reason = "crate::__crate_name! is intended")]
 macro_rules! doclink· {
     /* Rustdoc item filenames */
 
@@ -86,18 +85,18 @@ macro_rules! doclink· {
             "/", $crate::doclink![@item_file $kind $item] $(, $jump_link)? ]
     };
     (
-     // https://…/{crate::__crate_name!()}/module_path/struct.Item.html
+     // https://…/{env!("CARGO_PKG_NAME")}/module_path/struct.Item.html
      custom_current_crate $module_path:expr,
      @item $kind:ident $item:ident $($jump_link:literal)?) => {
         ::core::concat![ $crate::doclink![custom_current_crate $module_path,],
             "/", $crate::doclink![@item_file $kind $item] $(, $jump_link)? ]
     };
     (
-     // https://…/{crate::__crate_name!()}/struct.Item.html
+     // https://…/{env!("CARGO_PKG_NAME")}/struct.Item.html
      custom_current_proc_crate
      @item $kind:ident $item:ident $($jump_link:literal)?) => {
         ::core::concat![ crate::_DOCLINK_CUSTOM_DOMAIN!(),
-            crate::__crate_name!(), "/latest/", crate::__crate_name!(), "/",
+            env!("CARGO_PKG_NAME"), "/latest/", env!("CARGO_PKG_NAME"), "/",
             $crate::doclink![@item_file $kind $item] $(, $jump_link)? ]
     };
     /* existing module/path links */
@@ -114,8 +113,8 @@ macro_rules! doclink· {
      // file://…/current_crate/item_path/index.html
      custom crate $item_path:literal
      $(@mod$($_m:lifetime)?)? $($jump_link:literal)?) => {
-        ::core::concat![ crate::_DOCLINK_CUSTOM_DOMAIN!(), crate::__crate_name!(),
-            env!("CARGO_CRATE_NAME"), "/latest/", crate::__crate_name!(), "/",
+        ::core::concat![ crate::_DOCLINK_CUSTOM_DOMAIN!(), env!("CARGO_PKG_NAME"),
+            env!("CARGO_CRATE_NAME"), "/latest/", env!("CARGO_CRATE_NAME"), "/",
             $item_path $(, $jump_link)? ]
     };
     (
@@ -135,20 +134,20 @@ macro_rules! doclink· {
             "/latest/", ::core::stringify!($crate_name), "/", $item_path $(, $jump_link)? ]
     };
     (
-     // https://…/{crate::__crate_name!()}/item_path
-     // file://…/{crate::__crate_name!()}/item_path/index.html
+     // https://…/{env!("CARGO_PKG_NAME")}/item_path
+     // file://…/{env!("CARGO_PKG_NAME")}/item_path/index.html
      custom_current_crate $item_path:expr,
      $(@mod$($_m:lifetime)?)? $($jump_link:literal)?) => {
-        ::core::concat![ crate::_DOCLINK_CUSTOM_DOMAIN!(), crate::__crate_name!(),
-            "/latest/", crate::__crate_name!(), "/", $item_path $(, $jump_link)? ]
+        ::core::concat![ crate::_DOCLINK_CUSTOM_DOMAIN!(), env!("CARGO_PKG_NAME"),
+            "/latest/", env!("CARGO_PKG_NAME"), "/", $item_path $(, $jump_link)? ]
     };
     (
-     // https://…/{crate::__crate_name!()}/
-     // file://…/{crate::__crate_name!()}/index.html
+     // https://…/{env!("CARGO_PKG_NAME")}/
+     // file://…/{env!("CARGO_PKG_NAME")}/index.html
      custom_current_proc_crate
      $(@mod$($_m:lifetime)?)? $($jump_link:literal)?) => {
-        ::core::concat![ crate::_DOCLINK_CUSTOM_DOMAIN!(), crate::__crate_name!(),
-            "/latest/", crate::__crate_name!(), "/" $(, $jump_link)? ]
+        ::core::concat![ crate::_DOCLINK_CUSTOM_DOMAIN!(), env!("CARGO_PKG_NAME"),
+            "/latest/", env!("CARGO_PKG_NAME"), "/" $(, $jump_link)? ]
     };
     /* item links */
     (
@@ -182,18 +181,18 @@ macro_rules! doclink· {
             $crate::doclink![@item_file $kind $item] $(, $jump_link)? ]
     };
     (
-     // https://docs.rs/…/{crate::__crate_name!()}/module_path/struct.Item.html
+     // https://docs.rs/…/{env!("CARGO_PKG_NAME")}/module_path/struct.Item.html
      current_crate $module_path:literal
      @item $kind:ident $item:ident $($jump_link:literal)?) => {
         ::core::concat![ $crate::doclink![current_crate $module_path], "/",
             $crate::doclink![@item_file $kind $item] $(, $jump_link)? ]
     };
     (
-     // https://docs.rs/…/{crate::__crate_name!()}/struct.Item.html
+     // https://docs.rs/…/{env!("CARGO_PKG_NAME")}/struct.Item.html
      current_proc_crate
      @item $kind:ident $item:ident $($jump_link:literal)?) => {
-        ::core::concat![ "https://docs.rs/", crate::__crate_name!(), "/latest/",
-            crate::__crate_name!(), "/", $crate::doclink![@item_file $kind $item] $(, $jump_link)? ]
+        ::core::concat![ "https://docs.rs/", env!("CARGO_PKG_NAME"), "/latest/",
+            env!("CARGO_PKG_NAME"), "/", $crate::doclink![@item_file $kind $item] $(, $jump_link)? ]
     };
     /* existing module/path links */
     (
@@ -209,8 +208,8 @@ macro_rules! doclink· {
      // file://…/current_crate/item_path/index.html
      crate $item_path:literal
      $(@mod$($_m:lifetime)?)? $($jump_link:literal)?) => {
-        ::core::concat![ "https://docs.rs/", crate::__crate_name!(), env!("CARGO_CRATE_NAME"),
-            "/latest/", crate::__crate_name!(), "/", $item_path $(, $jump_link)? ]
+        ::core::concat![ "https://docs.rs/", env!("CARGO_PKG_NAME"), env!("CARGO_CRATE_NAME"),
+            "/latest/", env!("CARGO_CRATE_NAME"), "/", $item_path $(, $jump_link)? ]
     };
     (
      // [anchor]: https://…/crate_name/item_path
@@ -229,27 +228,26 @@ macro_rules! doclink· {
             ::core::stringify!($crate_name), "/", $item_path $(, $jump_link)? ]
     };
     (
-     // https://…/{crate::__crate_name!()}/item_path
-     // file://…/{crate::__crate_name!()}/item_path/index.html
+     // https://…/{env!("CARGO_PKG_NAME")}/item_path
+     // file://…/{env!("CARGO_PKG_NAME")}/item_path/index.html
      current_crate $item_path:literal
      $(@mod$($_m:lifetime)?)? $($jump_link:literal)?) => {
-        ::core::concat![ "https://docs.rs/", crate::__crate_name!(), "/latest/",
-            crate::__crate_name!(), "/", $item_path $(, $jump_link)? ]
+        ::core::concat![ "https://docs.rs/", env!("CARGO_PKG_NAME"), "/latest/",
+            env!("CARGO_PKG_NAME"), "/", $item_path $(, $jump_link)? ]
     };
     (
-     // https://…/{crate::__crate_name!()}/
-     // file://…/{crate::__crate_name!()}/index.html
+     // https://…/{env!("CARGO_PKG_NAME")}/
+     // file://…/{env!("CARGO_PKG_NAME")}/index.html
      current_proc_crate
      $(@mod$($_m:lifetime)?)? $($jump_link:literal)?) => {
-        ::core::concat![ "https://docs.rs/", crate::__crate_name!(), "/latest/",
-            crate::__crate_name!(), "/" $(, $jump_link)? ]
+        ::core::concat![ "https://docs.rs/", env!("CARGO_PKG_NAME"), "/latest/",
+            env!("CARGO_PKG_NAME"), "/" $(, $jump_link)? ]
     };
 }
 
 /// Helps doc-linking items in downstream crates.
 /// (*local* version)
 #[cfg(not(feature = "__publish"))]
-#[allow(clippy::crate_in_macro_def, reason = "crate::__crate_name! is intended")]
 macro_rules! doclink· {
     /* Rustdoc item filenames */
 
@@ -322,7 +320,7 @@ macro_rules! doclink· {
     (
      custom_current_proc_crate
      @item $kind:ident $item:ident $($jump_link:literal)?) => {
-        ::core::concat![ "file://", env!("CARGO_TARGET_DIR"), "doc/", crate::__crate_name!(),
+        ::core::concat![ "file://", env!("CARGO_TARGET_DIR"), "doc/", env!("CARGO_PKG_NAME"),
             "/", $crate::doclink![@item_file $kind $item] $(, $jump_link)? ]
     };
     /* existing module/path links */
@@ -356,17 +354,17 @@ macro_rules! doclink· {
             $(, "/index.html"$($_m)?)? $(, $jump_link)? ]
     };
     (
-     // file://…/{crate::__crate_name!()}/item_path/index.html
+     // file://…/{env!("CARGO_PKG_NAME")}/item_path/index.html
      custom_current_crate $item_path:expr,
      $(@mod$($_m:lifetime)?)? $($jump_link:literal)?) => {
-        ::core::concat![ "file://", env!("CARGO_TARGET_DIR"), "doc/", crate::__crate_name!(),
+        ::core::concat![ "file://", env!("CARGO_TARGET_DIR"), "doc/", env!("CARGO_PKG_NAME"),
             "/", $item_path $(, "/index.html"$($_m)?)? $(, $jump_link)? ]
     };
     (
-     // file://…/{crate::__crate_name!()}/index.html
+     // file://…/{env!("CARGO_PKG_NAME")}/index.html
      custom_current_proc_crate
      $(@mod$($_m:lifetime)?)? $($jump_link:literal)?) => {
-        ::core::concat![ "file://", env!("CARGO_TARGET_DIR"), "doc/", crate::__crate_name!()
+        ::core::concat![ "file://", env!("CARGO_TARGET_DIR"), "doc/", env!("CARGO_PKG_NAME")
             $(, "/index.html"$($_m)?)? $(, $jump_link)? ]
     };
     /* item links */
@@ -405,7 +403,7 @@ macro_rules! doclink· {
     (
      current_proc_crate
      @item $kind:ident $item:ident $($jump_link:literal)?) => {
-        ::core::concat![ "file://", env!("CARGO_TARGET_DIR"), "doc/", crate::__crate_name!(), "/",
+        ::core::concat![ "file://", env!("CARGO_TARGET_DIR"), "doc/", env!("CARGO_PKG_NAME"), "/",
             $crate::doclink![@item_file $kind $item] $(, $jump_link)? ]
     };
     /* existing module/path links */
@@ -439,17 +437,17 @@ macro_rules! doclink· {
             $(, "/index.html"$($_m)?)? $(, $jump_link)? ]
     };
     (
-     // file://…/{crate::__crate_name!()}/item_path/index.html
+     // file://…/{env!("CARGO_PKG_NAME")}/item_path/index.html
      current_crate $item_path:literal
      $(@mod$($_m:lifetime)?)? $($jump_link:literal)?) => {
-        ::core::concat![ "file://", env!("CARGO_TARGET_DIR"), "doc/", crate::__crate_name!(),
+        ::core::concat![ "file://", env!("CARGO_TARGET_DIR"), "doc/", env!("CARGO_PKG_NAME"),
             "/", $item_path $(, "/index.html"$($_m)?)? $(, $jump_link)? ]
     };
     (
-     // file://…/{crate::__crate_name!()}/index.html
+     // file://…/{env!("CARGO_PKG_NAME")}/index.html
      current_proc_crate
      $(@mod$($_m:lifetime)?)? $($jump_link:literal)?) => {
-        ::core::concat![ "file://", env!("CARGO_TARGET_DIR"), "doc/", crate::__crate_name!()
+        ::core::concat![ "file://", env!("CARGO_TARGET_DIR"), "doc/", env!("CARGO_PKG_NAME")
             $(, "/index.html"$($_m)?)? $(, $jump_link)? ] };
 }